add get_lcps and fix multiple results generated during iteration
There was a bug in the iteration logic which meant that during
back-tracking in composing results a shorter key would appear in
iteration results multiple times if it had multiple longer key entries
branching from the same stem.

This fix prevents that and improves performance by up to 10% by
eliminating this redunant (and incorrect) processing.

Add a get_lcps() function to find the list of valid "longest common
prefixes" for a trie.